requirements.txt
langchain==0.0.350
openai==1.12.0
chromadb==0.3.22
python-dotenv==1.0.0
PyPDF2==3.0.1
gradio==4.4.0
tiktoken==0.5.1 